feat: update ui style
